pyproject.toml
[project]
name = "kostenerstattung"
version = "0.1.2"
description = ""
authors = [
    {name = "kmille", email = "github@androidloves.me"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "bootstrap-flask (>=2.5.0,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "wtforms-alchemy (>=0.19.1,<0.20.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "zammad-py (>=3.2.0,<4.0.0)",
    "schwifty (>=2026.1.0,<2027.0.0)",
    "py-epc-qr @ git+https://github.com/kmille/py-epc-qr.git@0.1.5",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "kostenerstattung"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kostenerstattung = 'kostenerstattung.cli:main'
